Tile data structures moved to ligic package
