+++ src/logic/TileTypeException.java
@@ -1,4 +1,4 @@
-package tiles;
+package logic;
 
 public class TileTypeException extends RuntimeException{
     public  TileTypeException(String message) {
